Update bf-hit-rate and add bf-write test to count number of writes to buffer cache
Return unsigned instead of unsigned long long in syscalls

+++ pintos/pintos/src/lib/user/syscall.h
@@ -51,7 +51,7 @@ int inumber (int fd);
 void* sbrk (intptr_t increment);
 
 /* Buffer cache tests */
-unsigned long long block_read_count (void);
-unsigned long long block_write_count (void);
+unsigned block_read_count (void);
+unsigned block_write_count (void);
 
 #endif /* lib/user/syscall.h */
